requirements.txt
azure-ai-textanalytics
openai
